feat: CurationFeed Domain Test
